core: unify layout geometry and shrink interface (03)
- layout: collapse split_horizontal/vertical_n into single split_n
  with SplitDirection axis param; ratios Vec<f32> -> Vec<f64> with
  debug_assert on len mismatch vs children; no f32->f64 cast churn.
  Shrink public interface to 3 fns (calculate_layout, window_at_point,
  bar_strip_rect); inset_rect / screen_local_to_global /
  subtract_strip are pub(crate). Add subtract_strip (exact equality,
  no EPSILON 0.5) to replace workspace remove_edge_strip dance;
  3 new unit tests for strip subtraction.
- tree: NodeData::Split ratios Vec<f32> -> Vec<f64>.
- workspace: add is_widescreen helper, unify next_direction (single
  default + flip on parent split) removing 4× widescreen duplication;
  usable_rect now calls layout::subtract_strip (no EPSILON, same
  integer-derived geometry); rebalance uses f64. 72 core tests pass.

+++ pengwm-core/src/tree.rs
@@ -39,7 +39,7 @@ pub enum NodeData {
     },
     Split {
         direction: SplitDirection,
-        ratios: Vec<f32>,
+        ratios: Vec<f64>,
     },
 }
